// generated/orval/fetch/google-health-api/models/activeZoneMinutesHeartRateZone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Required. Heart rate zone in which the active zone minutes have been earned, in the given time interval.
 */
export type ActiveZoneMinutesHeartRateZone = typeof ActiveZoneMinutesHeartRateZone[keyof typeof ActiveZoneMinutesHeartRateZone];


export const ActiveZoneMinutesHeartRateZone = {
  HEART_RATE_ZONE_UNSPECIFIED: 'HEART_RATE_ZONE_UNSPECIFIED',
  FAT_BURN: 'FAT_BURN',
  CARDIO: 'CARDIO',
  PEAK: 'PEAK',
} as const;