// generated/orval/fetch/google-health-api/models/observationTimeInterval.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { CivilDateTime } from './civilDateTime';

/**
 * Represents a time interval of an observed data point.
 */
export interface ObservationTimeInterval {
  /** Output only. Observed interval end time in civil time in the timezone the subject is in at the end of the observed interval */
  readonly civilEndTime?: CivilDateTime;
  /** Output only. Observed interval start time in civil time in the timezone the subject is in at the start of the observed interval */
  readonly civilStartTime?: CivilDateTime;
  /** Required. Observed interval end time. */
  endTime?: string;
  /** Required. The offset of the user's local time at the end of the observation relative to the Coordinated Universal Time (UTC). */
  endUtcOffset?: string;
  /** Required. Observed interval start time. */
  startTime?: string;
  /** Required. The offset of the user's local time at the start of the observation relative to the Coordinated Universal Time (UTC). */
  startUtcOffset?: string;
}